+++ packages/example-antd-management-fast-framework/src/customSpecialComponents/FunctionExtra/ScoreProductStockChangeType/index.jsx
@@ -11,6 +11,7 @@ import {
   unlimitedWithStringFlag,
 } from 'antd-management-fast-common';
 import {
+  buildDropdownMenu,
   buildFlexRadio,
   buildFlexSelect,
   buildFormRadio,
@@ -88,6 +89,7 @@ export function renderSearchScoreProductStockChangeTypeSelect({
 }
 
 export function renderCustomScoreProductStockChangeTypeSelect({
+  style = {},
   label = '变更类型',
   separator = ':',
   size = 'middle',
@@ -95,6 +97,7 @@ export function renderCustomScoreProductStockChangeTypeSelect({
   innerProps: innerProperties = null,
 }) {
   return buildFlexSelect({
+    style,
     label,
     defaultValue: null,
     separator,
@@ -104,6 +107,26 @@ export function renderCustomScoreProductStockChangeTypeSelect({
     renderItem: null,
     onChange: onChangeCallback,
     innerProps: innerProperties,
+  });
+}
+
+export function renderScoreProductStockChangeTypeDropDown({
+  label = '变更类型',
+  placement = 'bottomRight',
+  icon = null,
+  size = 'middle',
+  onClick: onClickCallback,
+  innerProps: innerProperties = null,
+}) {
+  return buildDropdownMenu({
+    label,
+    placement,
+    icon,
+    size,
+    list: refitScoreProductStockChangeTypeList({ withUnlimited: false }),
+    dataConvert: null,
+    onClick: onClickCallback,
+    innerProps: innerProperties || null,
   });
 }